Bug fix (temp = 0 +- 1)
When the temp got to zero, the code
thought that it had invalid data.
Because of that, it wouldn't display a 0
temp, the color of the temp was displayed as white for -1 and 1 after it had been 0, and it wouldn't log -1 as the min temp (or 1 as the max temp) for the day.